refined tests, reorganized project structure
