+++ src/project_handler_factory.py
@@ -1,8 +1,11 @@
 from enum import Enum
-from src.services.project_operations import InMemProjectHandler
+
+from src.services.inmem_project_handler import InMemProjectHandler
+
 
 class ProjectHandlerType(Enum):
     IN_MEMORY = 1
+
 
 def createHandler(type: ProjectHandlerType) -> object:
     match type:
