+++ src/project_handler_interface.py
@@ -1,11 +1,20 @@
 from abc import ABC, abstractmethod
 
+
 class ProjectHandlerInterface(ABC):
-    
+
     @abstractmethod
-    def create(self, name: str, ownerId: int, description: str, logo: str = None, documents: str = None, contributors: list[int] = None) -> None:
+    def create(
+        self,
+        name: str,
+        ownerId: int,
+        description: str,
+        logo: str = None,
+        documents: str = None,
+        contributors: list[int] = None,
+    ) -> None:
         raise NotImplementedError
-    
+
     @abstractmethod
     def get_all(self):
         raise NotImplementedError
@@ -13,7 +22,7 @@ class ProjectHandlerInterface(ABC):
     @abstractmethod
     def get(self, projectId: int) -> object:
         raise NotImplementedError
-    
+
     @abstractmethod
     def update_info(self, projectId: int, attributesToUpdate: dict) -> None:
         raise NotImplementedError
